Reorganizo el código de indexación
    refs:#142
