+++ series_tiempo_ar_api/apps/api/tests/reader_tests.py
@@ -10,7 +10,7 @@ from series_tiempo_ar.search import get_time_series_distributions
 
 from series_tiempo_ar_api.apps.api.indexing.database_loader import \
     DatabaseLoader
-from series_tiempo_ar_api.apps.api.indexing.indexer import DistributionIndexer
+from series_tiempo_ar_api.apps.api.indexing.distribution_indexer import DistributionIndexer
 from series_tiempo_ar_api.apps.api.indexing.scraping import Scraper
 from series_tiempo_ar_api.apps.api.models import Dataset, Distribution, Field, Catalog
 from series_tiempo_ar_api.apps.api.query.elastic import ElasticInstance
