+++ series_tiempo_ar_api/apps/management/management/commands/read_datajson.py
@@ -4,7 +4,7 @@ import logging
 from django.core.management import BaseCommand
 from pydatajson import DataJson
 
-from series_tiempo_ar_api.apps.api.indexing.catalog_reader import index_catalog
+from series_tiempo_ar_api.apps.api.indexing import catalog_reader
 from series_tiempo_ar_api.apps.management.models import Node
 
 logger = logging.getLogger(__name__)
@@ -20,8 +20,9 @@ class Command(BaseCommand):
             catalog_url = node.catalog_url
             try:
                 catalog = DataJson(catalog_url)
-            except (IOError, ValueError) as e:
+            except (IOError, ValueError, AssertionError) as e:
+                self.stderr.write(READ_ERROR % (catalog_id, e))
                 logging.warn(READ_ERROR, catalog_id, e)
                 continue
 
-            index_catalog(catalog, catalog_id)
+            catalog_reader.index_catalog(catalog, catalog_id)
